boards/arm/stm32f7/stm32f746g-disco: add SAI driver

* ignore TAGS

* arch/arm/src/stm32l4/stm32l4_sai.c: fix type error

Approved-by: Gregory Nutt <gnutt@nuttx.org>

+++ arch/arm/src/stm32l4/stm32l4_sai.c
@@ -490,7 +490,7 @@ static void sai_exclsem_take(struct stm32l4_sai_s *priv)
 
 static void sai_mckdivider(struct stm32l4_sai_s *priv)
 {
-  uint8_t mckdiv;
+  uint32_t mckdiv;
 
   DEBUGASSERT(priv && priv->samplerate > 0 && priv->frequency > 0);
